Add initial Slack app
Adds the ability for a Slack workspace to install Hiro
via OAuth.

Adds a Slack app that allows workspaces to add
customized triggers.

Adds an example trigger action for current use in the Rails Slack
workspace which monitors the #work-offers channel, and responds
with a reminder to include as much relevant information as possible.

Adds support for self-destruction reactions to bot-posted
messages.

Brings in support for a custom fork of the crystal-db adapter
to add support for TTLs on the connections in the DB Pool.